backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning and NLP
openai-whisper==20231117
spacy==3.7.2
dateparser==1.2.0

# Audio processing
ffmpeg-python==0.2.0

# Data handling
pydantic==2.5.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0 